Handle URLs better in github deployment
Locally the /static folder's handled well, but in deployment there's a
subfolder (/blog/) that messes up the images and videos.

This adds some possibly questionable (no clue what I'm doing) shortcodes
that use the relURL in the html src= bits to actually find the
appropriate media in deployment.

Hopefully this works, as to test I need to deploy.. so let's ship it and
see!